Updated Forth environment to kForth-Win32 v1.6.6; updated XYPLOT-Win32 version #.
 Changes to be committed:
	modified:   bin/xypw32.exe
	new file:   doc/dpans94.pdf
	new file:   doc/kForth-32 Manual.pdf
	modified:   modules/ans-words.4th
	modified:   modules/files.4th
	modified:   modules/strings.4th
	modified:   modules/struct.4th
	modified:   src/CPlotWindow.cpp
	modified:   src/ForthCompiler.cpp
	modified:   src/ForthCompiler.h
	modified:   src/ForthVM.cpp
	modified:   src/ForthVM.h
	modified:   src/ForthWords.h
	modified:   src/fbc.h
	modified:   src/vm32.asm
	modified:   src/vm32.obj
	modified:   src/vmc.c
	modified:   src/xydlgs.dlg
	modified:   src/xyplot.cpp
	modified:   src/xyplot.h

+++ src/xyplot.h
@@ -29,7 +29,7 @@ int AddToHeader (char*, char*, bool);
 void LoadInitializationFile();
 void InitForthInterface ();
 int LoadForthFile(char*);
-int ExecuteForthExpression (char*, ostringstream*, int*, int*);
+int ExecuteForthExpression (char*, ostringstream*, int*, long int*);
 COLORREF LookupColor(char*);
 
 // The following functions provide the interface between
